Updating asteroid-spawning-system branch (#7)
* Started work on asteroid spawning system

Asteroids spawn correctly however position and frequency needs to be changed for the game. Currently, the game spawns between three and six random asteroids on the same position every frame.

* Final session commit

* Merging asset DB changes

* Pulling in old branches for new branch structure (#6)

* Started work on asteroid spawning system

Asteroids spawn correctly however position and frequency needs to be changed for the game. Currently, the game spawns between three and six random asteroids on the same position every frame.

* Final session commit

* Merging asset DB changes

* Preparing for merge