Removing unnecessary files (#9)
* Started work on asteroid spawning system

Asteroids spawn correctly however position and frequency needs to be changed for the game. Currently, the game spawns between three and six random asteroids on the same position every frame.

* Final session commit

* Merging asset DB changes

* Pulling in old branches for new branch structure (#6)

* Started work on asteroid spawning system

Asteroids spawn correctly however position and frequency needs to be changed for the game. Currently, the game spawns between three and six random asteroids on the same position every frame.

* Final session commit

* Merging asset DB changes

* Preparing for merge

* Update Unity.gitignore

* Delete Library directory

* Delete Logs directory

* Delete .idea/.idea.TopDownFramework/.idea directory

* Delete .vscode directory

* Delete TopDownFramework.sln.DotSettings.user

* Delete Unity.gitignore

* Delete Temp directory

* Delete obj/Debug directory

* Delete UserSettings directory

* Delete .gitattributes